# public/dist/swagger.yaml
---
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Viagens"
host: "localhost:3000"
basePath: "/"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
  /viagens:
    get:
      description: "Returns all travels from the system that the user has access to"
      operation_id: "findtravels"
      produces:
      - "application/json"
      - "application/xml"
      - "text/xml"
      - "text/html"
      - name: "Pais"
        in: "query"
        description: "maximum number of results to return"
        required: false
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "travel response"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/travel"
        default:
          schema:
            $ref: "#/definitions/errorModel"
    post:
      description: "Creates a new travel in the store.  Duplicates are allowed"
      operation_id: "addtravel"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "Pais"
        description: "travel to add to the store"
        required: true
        schema:
          $ref: "#/definitions/newtravel"
      responses:
        200:
          description: "travel response"
          schema:
            $ref: "#/definitions/travel"
        default:
          schema:
            $ref: "#/definitions/errorModel"
  /viagens/{_id}:
    get:
      description: "Returns a user based on a single _id, if the user does not have\
        \ access to the travel"
      operation_id: "findtravelBy_id"
      produces:
      - "application/json"
      - "application/xml"
      - "text/xml"
      - "text/html"
      parameters:
      - name: "_id"
        in: "path"
        description: "_id of travel to fetch"
        required: true
        type: "object"
      responses:
        200:
          description: "travel response"
          schema:
            $ref: "#/definitions/travel"
        default:
          schema:
            $ref: "#/definitions/errorModel"
    delete:
      description: "deletes a single travel based on the _id supplied"
      operation_id: "deletetravel"
      parameters:
      - name: "_id"
        in: "path"
        description: "_id of travel to delete"
        required: true
        type: "object"
      responses:
        204:
          description: "travel deleted"
        default:
          schema:
            $ref: "#/definitions/errorModel"
definitions:
  travel:
    type: "object"
    required:
    - "_id"
    - "Pais"
    properties:
      Pais:
        type: "string"
      Cidade:
        type: "string"
      Descricao:
        type: "string"
      Localizacao:
        type: "number"
        format: "point"
      Visivel:
        type: "string"
        format: "Boolean"
      Chegada:
        type: "string"
        format: "date"
      Partida:
        type: "string"
        format: "date"
      idUtilizador:
        type: "string"
  newtravel:
    type: "object"
    required:
    - "Pais"
    properties:
      Pais:
        type: "string"
      Cidade:
        type: "string"
      Descricao:
        type: "string"
      Localizacao:
        type: "number"
        format: "point"
      Visivel:
        type: "string"
        format: "Boolean"
      Chegada:
        type: "string"
        format: "date"
      Partida:
        type: "string"
        format: "date"
      idUtilizador:
        type: "string"
  errorModel:
    type: "object"
    required:
    - "code"
    - "message"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"